+++ src/onefinity/onefinity.py
@@ -6,7 +6,8 @@ from onefinity import random_words
 def main():
     print("Welcome to OneFinity - A number guessing game!")
     print("1. Number Guesser")
-    print("2. Word Guesser")
+    print("2. Number guesser (Computer)")
+    print("3. Word Guesser")
     choice = int(input("Enter choice: "))
 
     if choice == 1:
@@ -14,10 +15,14 @@ def main():
         number_guesser()
     elif choice == 2:
         os.system('clear')
+        number_to_guess()
+    elif choice == 3:
+        os.system('clear')
         word_guesser()
     else:
         print("Incorrect option entered!")
         exit
+
 
 def number_guesser():
     print("The computer will choose a number between 1 - 10")
@@ -50,6 +55,7 @@ def word_guesser():
     print("1. Random words")
     print("2. Animals")
     print("3. Cities")
+    print("4. Countries")
     choice = int(input("Enter: "))
     
     if choice == 1:
@@ -58,6 +64,8 @@ def word_guesser():
         chosen_word = random.choice(random_words.animals)
     elif choice == 3:
         chosen_word = random.choice(random_words.cities)
+    elif choice == 4:
+        chosen_word = random.choice(random_words.countries)
     else:
         print("Invalid choice!")
         exit
@@ -69,4 +77,35 @@ def word_guesser():
         print("Right, you won!")
     else:
         print("That is not right.")
+    
+    
+def number_to_guess():
+    print("Give the minimum and maximum value for the range and choose a number")
+    print("The computer will try to guess it")
+    
+    min_number = int(input("Enter minimum number: "))
+    max_number = int(input("Enter maximum number: "))
+    
+    computer_guess = random.randrange(min_number, max_number)
+    print(f"Is the number {computer_guess}?")
+    
+    choice = input("Enter (higher / lower / yes) : ")
+    
+    while choice != "yes":
+        if choice == "higher":
+            max_number = computer_guess
+            computer_guess = random.randrange(min_number, max_number)
+            print(f"Is it {computer_guess}?")
+            choice = input("Enter (higher / lower / yes) : ")
+        elif choice == "lower":
+            min_number = computer_guess
+            computer_guess = random.randrange(min_number, max_number)
+            print(f"Is it {computer_guess}?")
+            choice = input("Enter (higher / lower / yes) : ")
+        else:
+            break
+        
+    print("I got it right. Thank you for playing.")
+    
+