+++ src/onefinity/random_words.py
@@ -79,4 +79,26 @@ cities = [
     "wroclaw","xian","yangon","yerevan","zagreb","zanzibar city","zaragoza","zhengzhou","zibo","zurich"
 ]
 
+countries = [
+    "afghanistan","albania","algeria","andorra","angola","antigua_and_barbuda","argentina","armenia","australia","austria",
+    "azerbaijan","bahamas","bahrain","bangladesh","barbados","belarus","belgium","belize","benin","bhutan",
+    "bolivia","bosnia_and_herzegovina","botswana","brazil","brunei","bulgaria","burkina faso","burundi","cambodia","cameroon",
+    "canada","cape_verde","central_african_republic","chad","chile","china","colombia","comoros","congo","costa rica",
+    "croatia","cuba","cyprus","czechia","denmark","djibouti","dominica","dominican_republic","ecuador","egypt",
+    "el_salvador","equatorial_guinea","eritrea","estonia","eswatini","ethiopia","fiji","finland","france","gabon",
+    "gambia","georgia","germany","ghana","greece","grenada","guatemala","guinea","guinea_bissau","guyana",
+    "haiti","honduras","hungary","iceland","india","indonesia","iran","iraq","ireland","israel",
+    "italy","ivory_coast","jamaica","japan","jordan","kazakhstan","kenya","kiribati","kosovo","kuwait",
+    "kyrgyzstan","laos","latvia","lebanon","lesotho","liberia","libya","liechtenstein","lithuania","luxembourg",
+    "madagascar","malawi","malaysia","maldives","mali","malta","marshall_islands","mauritania","mauritius","mexico",
+    "micronesia","moldova","monaco","mongolia","montenegro","morocco","mozambique","myanmar","namibia","nauru",
+    "nepal","netherlands","new zealand","nicaragua","niger","nigeria","north_korea","north_macedonia","norway","oman",
+    "pakistan","palau","panama","papua new guinea","paraguay","peru","philippines","poland","portugal","qatar",
+    "romania","russia","rwanda","saint kitts and nevis","saint_lucia","saint_vincent_and_the_grenadines","samoa","san marino","sao tome and principe","saudi arabia",
+    "senegal","serbia","seychelles","sierra_leone","singapore","slovakia","slovenia","solomon_islands","somalia","south_africa",
+    "south_korea","south_sudan","spain","sri_lanka","sudan","suriname","sweden","switzerland","syria","taiwan",
+    "tajikistan","tanzania","thailand","timor leste","togo","tonga","trinidad_and_tobago","tunisia","turkey","turkmenistan",
+    "tuvalu","uganda","ukraine","united arab emirates","united kingdom","united states","uruguay","uzbekistan","vanuatu","vatican city",
+    "venezuela","vietnam","yemen","zambia","zimbabwe"
+]