+++ CHANGELOG.md
@@ -1,7 +1,10 @@
-# Version 0.5 (Current)
+# Version 0.6 (Current)
+- Added new game option -- Number Guesser (Computer)
+- Added one new words list -- countries
+# Version 0.5
 - Added two new words list -- animals and cities
 - Made slight adjustments to the code
-# Version 0.4 (Current)
+# Version 0.4
 - Added 200 words for the word guesser game
 - Added attempt counter and points calculator for number guesser
 # Version 0.3